requirements.txt
absl-py==0.11.0
appdirs==1.4.3
astunparse==1.6.3
attrs==20.3.0
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
Click==7.0
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.3.2
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.63.0
flatbuffers==1.12
gast==0.3.3
google-auth==1.26.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
gplaycli==3.29
grpcio==1.32.0
gunicorn==20.0.4
h11==0.8.1
h5py==2.10.0
html5lib==1.0.1
httptools==0.0.13
idna==2.8
importlib-metadata==3.3.0
iniconfig==1.1.1
ipaddr==2.2.0
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lockfile==0.12.2
lxml==4.6.2
Markdown==3.3.3
matlink-gpapi==0.4.4.5
matplotlib==3.3.4
more-itertools==8.6.0
msgpack==0.6.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.8
pep517==0.8.2
Pillow==8.1.2
pluggy==0.13.1
prettytable==2.1.0
progress==1.5
protobuf==3.14.0
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyaxmlparser==0.3.24
pycparser==2.20
pydantic==1.2
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-multipart==0.0.5
pytoml==0.1.21
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
starlette==0.13.6
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
